refactor(core): centralize participant session runtime
