+++ package.json
@@ -27,5 +27,5 @@
     "typescript": "^5",
     "ultracite": "7.0.7"
   },
-  "packageManager": "bun@1.3.10"
+  "packageManager": "bun@1.3.13"
 }
