+++ packages/cli/src/utils/management-api.ts
@@ -228,10 +228,11 @@ export async function* watchRoomEvents(
 
   const reader = response.body.getReader();
   const decoder = new TextDecoder();
+  type ReaderChunk = Awaited<ReturnType<typeof reader.read>>;
   let buffer = "";
 
   while (true) {
-    let chunk;
+    let chunk: ReaderChunk;
     try {
       chunk = await reader.read();
     } catch (error) {
