+++ packages/core/src/room.ts
@@ -255,7 +255,9 @@ function updateLastSeen(roomId: string, participantId: string): boolean {
   participant.lastSeen = Date.now();
   participant.updatedAt = participant.lastSeen;
   if (participant.status !== "busy") {
-    participant.status = participant.connection.connected ? "online" : "offline";
+    participant.status = participant.connection.connected
+      ? "online"
+      : "offline";
   }
   return true;
 }
