+++ packages/sdk/AGENTS.md
@@ -18,7 +18,7 @@ Arquivos correspondentes:
 - `src/provider.ts` — `createGambi()` (AI SDK v5 provider), `gambi.any()/model()/participant()`, `gambi.openResponses.*`, `gambi.chatCompletions.*`, `gambi.listModels()`, `gambi.listParticipants()`, `gambi.baseURL`.
 - `src/client.ts` — `createClient()`, namespaces (`rooms`, `participants`, `events`), `ClientError`.
 - `src/rooms.ts`, `src/participants.ts`, `src/hub.ts` — implementações dos namespaces e chamadas ao management plane.
-- `src/participant-session.ts` — `createParticipantSession()` e `ParticipantSession` (close reasons: `"closed"`, `"heartbeat_failed"`, `"tunnel_closed"`).
+- `src/participant-session.ts` — reexport de `createParticipantSession()` e `ParticipantSession` implementados em `@gambi/core/participant-session` (close reasons: `"closed"`, `"heartbeat_failed"`, `"tunnel_closed"`).
 - `src/discovery.ts` — `discoverHubs()`, `discoverRooms()`, `resolveGambiTarget()`, `DiscoveryError`.
 - `src/protocol.ts`, `src/protocol-adapters.ts` — seleção Responses vs Chat Completions.
 - `src/types.ts` — tipos públicos reexportados (inclui `HEALTH_CHECK_INTERVAL` para runtimes custom).
@@ -36,7 +36,7 @@ Qualquer mudança aqui atualiza `apps/docs/src/content/docs/reference/sdk.md` no
 ## Regras de design
 
 - **Explicitness sobre mágica**: `createGambi()` e `createClient()` são factories **explícitas**. Não adicionar discovery implícito nelas — quem quiser descoberta usa `resolveGambiTarget()` primeiro e passa o resultado.
-- **`createParticipantSession()` é o runtime canônico**: o CLI `gambi participant join` construiu em cima dela. Novos runtimes custom devem usar essa mesma função em vez de reimplementar tunelamento.
+- **`createParticipantSession()` é o runtime canônico**: a implementação vive em `@gambi/core/participant-session`, e a SDK reexporta essa função para usuários externos. Novos runtimes custom devem usar essa mesma função em vez de reimplementar tunelamento.
 - **Auth headers só saem no leg participante→provider**: `ParticipantAuthHeaders` passados a `createParticipantSession()` são aplicados apenas ao chamar o endpoint local do provider. Não enviar em nenhum outro lugar; o hub jamais os recebe.
 - **Heartbeat/ping em `HEALTH_CHECK_INTERVAL` (10 s)**: constante re-exportada de `types.ts`. Se construir um runtime custom sem `createParticipantSession()`, usar essa mesma cadência — a janela de offline do hub é `PARTICIPANT_TIMEOUT` (30 s).
 - **Management plane do core é fonte única**: a SDK não inventa rotas. Se precisar de algo novo, adicionar no `@gambi/core` primeiro.
