+++ packages/sdk/src/index.ts
@@ -1,5 +1,6 @@
 // Main SDK entry point - orchestrates all exports
 
+export type * from "@gambi/core/tunnel-protocol";
 // HTTP client for remote hubs
 export { ClientError, createClient, type GambiClient } from "./client.ts";
 // Optional discovery helpers for local-network apps
@@ -24,18 +25,17 @@ export type {
   ResponsesCreateRequest,
 } from "./hub.ts";
 export { hub } from "./hub.ts";
-export { participants } from "./participants.ts";
 export {
   createParticipantSession,
   type ParticipantSession,
   type ParticipantSessionCloseEvent,
   type ParticipantSessionOptions,
 } from "./participant-session.ts";
+export { participants } from "./participants.ts";
 // Protocol messages (re-exported from core)
 export type * from "./protocol.ts";
 export type * from "./protocol-adapters.ts";
 export * from "./protocol-adapters.ts";
-export type * from "@gambi/core/tunnel-protocol";
 // Existing AI SDK provider
 export {
   createGambi,
