+++ packages/sdk/src/types.ts
@@ -12,8 +12,8 @@ export type {
   ChatCompletionMessage,
   ChatCompletionMessageParam,
   GenerationConfig,
-  HubConfig,
   HeartbeatResult,
+  HubConfig,
   InputItemListParams,
   LlmMetrics,
   MachineSpecs,
@@ -27,10 +27,8 @@ export type {
   ParticipantInfo,
   ParticipantInfoInternal,
   ParticipantRegistration,
-  ParticipantSummary,
   ParticipantStatus,
-  RoomEvent,
-  RoomEventType,
+  ParticipantSummary,
   Response,
   ResponseCreateParams,
   ResponseCreateParamsNonStreaming,
@@ -42,6 +40,8 @@ export type {
   ResponseRetrieveParamsStreaming,
   ResponseStreamEvent,
   ResponseUsage,
+  RoomEvent,
+  RoomEventType,
   RoomInfo,
   RoomInfoPublic,
   RoomSummary,
@@ -58,8 +58,8 @@ export {
   ApiMeta as ApiMetaSchema,
   GenerationConfig as GenerationConfigSchema,
   HEALTH_CHECK_INTERVAL,
-  HubConfig as HubConfigSchema,
   HeartbeatResult as HeartbeatResultSchema,
+  HubConfig as HubConfigSchema,
   LlmMetrics as LlmMetricsSchema,
   MachineSpecs as MachineSpecsSchema,
   NetworkConfig as NetworkConfigSchema,
@@ -70,8 +70,8 @@ export {
   ParticipantInfo as ParticipantInfoSchema,
   ParticipantInfoInternal as ParticipantInfoInternalSchema,
   ParticipantRegistration as ParticipantRegistrationSchema,
-  ParticipantSummary as ParticipantSummarySchema,
   ParticipantStatus as ParticipantStatusSchema,
+  ParticipantSummary as ParticipantSummarySchema,
   RoomEvent as RoomEventSchema,
   RoomEventType as RoomEventTypeSchema,
   RoomInfoInternal as RoomInfoSchema,
